前端做了设备界面，后端梳理了整体逻辑，未测

+++ opsdev_monitor/src/main/java/org/alvin/opsdev/monitor/system/service/CheckMetricService.java
@@ -2,19 +2,21 @@ package org.alvin.opsdev.monitor.system.service;
 
 import org.alvin.opsdev.monitor.system.bean.CollectorTicket;
 import org.alvin.opsdev.monitor.system.bean.action.NotifyActionExecutor;
-import org.alvin.opsdev.monitor.system.service.cache.StatusCacheBean;
 import org.alvin.opsdev.monitor.system.bean.enums.AlertLevel;
 import org.alvin.opsdev.monitor.system.bean.enums.AlertStatus;
+import org.alvin.opsdev.monitor.system.bean.enums.MetricType;
 import org.alvin.opsdev.monitor.system.bean.enums.ObjectStatus;
 import org.alvin.opsdev.monitor.system.domain.*;
 import org.alvin.opsdev.monitor.system.service.cache.AlertCacheService;
 import org.alvin.opsdev.monitor.system.service.cache.DeviceMetricValueCacheService;
 import org.alvin.opsdev.monitor.system.service.cache.DeviceStatusCacheService;
+import org.alvin.opsdev.monitor.system.service.cache.StatusCacheBean;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.util.Assert;
 
 import java.util.List;
+import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * Created by tangzhichao on 2017/4/24.
@@ -47,27 +49,32 @@ public class CheckMetricService {
      * @param ticket
      */
     public void check(List<DeviceGroup> groups, List<Device> devices, CollectorTicket ticket) {
+        AtomicInteger count = new AtomicInteger(0);
         //先检查所有状态出错的，或者恢复的
-        checkStatus(groups, devices, ticket);
+        checkStatus(groups, devices, ticket,count);
         //检查指标是否有出错的，或者有恢复的
-        checkMetrics(groups, devices, ticket);
+        checkMetrics(groups, devices, ticket,count);
+        // 没有产生告警，直接退出
+        if(count.get() == 0){
+            return ;
+        }
         // TODO: 2017/4/25 如果有新的严重级别的问题，并且超过了三次，立即发送一封邮件
         this.notifyActionExecutor.checkAlert();
     }
 
     //_______________________________status__________________________________________________
-    private void checkStatus(List<DeviceGroup> groups, List<Device> devices, CollectorTicket ticket) {
+    private void checkStatus(List<DeviceGroup> groups, List<Device> devices, CollectorTicket ticket,AtomicInteger count) {
         groups.forEach(group -> {
             List<Device> devList = deviceService.findByGroupAndEnabled(group);
             Assert.notNull(devList, "checkStatus : device must not be null");
-            devList.forEach(device -> checkDevStatus(device, ticket));
+            devList.forEach(device -> checkDevStatus(device, ticket,count));
         });
-        devices.forEach(device -> checkDevStatus(device, ticket));
+        devices.forEach(device -> checkDevStatus(device, ticket,count));
     }
 
-    private void checkDevStatus(Device device, CollectorTicket ticket) {
+    private void checkDevStatus(Device device, CollectorTicket ticket,AtomicInteger count) {
         StatusCacheBean statusCacheBean = this.deviceStatusCacheService.get(device.getId());
-        ObjectStatus status = device.getStatus();
+        ObjectStatus status = statusCacheBean.getStatus();
         if (statusCacheBean == null) {
             if (status != ObjectStatus.SUCCESS) {
                 //状态缓存
@@ -81,6 +88,8 @@ public class CheckMetricService {
                 alert.setStatus(AlertStatus.ACTIVE);
                 //产生一个告警信息
                 this.alertService.save(alert);
+                //产生新的告警，需要计数器记录一下
+                count.incrementAndGet();
             }
             return;
         }
@@ -92,6 +101,8 @@ public class CheckMetricService {
             alert.setCount(1);
             alert.setTime(ticket.getTime());
             alert.setStatus(AlertStatus.ACTIVE);
+            //产生新的告警，需要计数器记录一下
+            count.incrementAndGet();
         }
         if (status == ObjectStatus.SUCCESS && status != statusCacheBean.getStatus()) {
             //产生改为清除，次数清0 ,清缓存
@@ -112,19 +123,25 @@ public class CheckMetricService {
     }
 
     //_______________________________metric__________________________________________________
-    private void checkMetrics(List<DeviceGroup> groups, List<Device> devices, CollectorTicket ticket) {
+    private void checkMetrics(List<DeviceGroup> groups, List<Device> devices, CollectorTicket ticket,AtomicInteger count) {
+        //设备组检查
         groups.forEach(group -> {
             List<Device> devList = this.deviceService.findByGroupAndEnabled(group);
             Assert.notNull(devList, "checkStatus : device must not be null");
-            devList.forEach(device -> checkDevMetric(device, ticket));
+            devList.forEach(device -> checkDevMetric(device, ticket,count));
         });
-        devices.forEach(device -> checkDevMetric(device, ticket));
+        //无组设备检查
+        devices.forEach(device -> checkDevMetric(device, ticket,count));
     }
 
-    private void checkDevMetric(Device device, CollectorTicket ticket) {
+    private void checkDevMetric(Device device, CollectorTicket ticket,AtomicInteger count) {
         List<Metric> metrics = this.metricService.findByObjectType(device.getCollectorType());
         metrics.forEach(metric -> {
             Threshold threshold = this.thresholdService.findByDeviceAndMetric(device, metric);
+            //属性值用来查看，不用来比较 ，比如cpu 个数,内存总量
+            if (metric.getMetricType().equals(MetricType.ATTR)) {
+                return;
+            }
             if (threshold == null) {
                 //Todo  清除缓存
                 this.alertCacheService.remove(device, metric);
@@ -135,13 +152,26 @@ public class CheckMetricService {
                 this.alertCacheService.remove(device, metric);
                 return;
             }
-            double value = this.deviceMetricValueCacheService.get(device, metric);
             Alert alert = this.alertCacheService.get(device, metric);
             if (alert == null) {
                 alert = new Alert();
                 alert.setStatus(AlertStatus.ACTIVE);
                 alert.setCount(0);
                 alert.setLevel(AlertLevel.NORMAL);
+                //产生新的告警，需要计数器记录一下
+                count.incrementAndGet();
+            }
+            Double value = this.deviceMetricValueCacheService.get(device, metric);
+            if(value == null){
+                //Todo 严重错误 没有获取到属性值
+                alert.setTime(ticket.getTime());
+                alert.setDevice(device);
+                alert.setMetric(metric);
+                alert.setLevel(AlertLevel.CRITICAL);
+                alert.setStatus(AlertStatus.UNKNOWN);
+                alert.setCount(alert.getCount() + 1);
+                this.alertCacheService.put(device, metric, alert);
+                return;
             }
             if (value >= threshold.getLimit()) {
                 //Todo 严重错误
